requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
slack_bolt==1.19.1
slack_sdk==3.31.0
soupsieve==2.5
urllib3==2.2.2
